Refactor CLI core commands
